+++ changelog.md
@@ -1,6 +1,9 @@
 # v0.2.3 (Upcoming)
+<<<<<<< HEAD
 * Add option to calculate density in the pore with a constant bin area.
 * Add option to specify a upper integration limit when calculating the mean density
+* New angle routine
+
 
 # v0.2.2
 * Routines can save and load results in hdf5 or obj format